added Main and Header component plus some styling
