+++ code/src/components/Header.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Header = () => {
+  return <h1 className="heading-one">&#128525; Happy thoughts &#128525;</h1>;
+};
+
+export default Header;
