--- code/src/components/LoadingItem.js
@@ -1,15 +1,0 @@
-import React from "react";
-
-const LoadingItem = () => {
-
-    return (
-        <div className="loading-overlay">
-
-            <div  className="loading-spinner" />
-
-
-        </div>
-    );
-}
-
-export default LoadingItem;
