+++ code/src/components/Spinner.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Spinner = () => {
+  return (
+    <div className="loading-overlay">
+      <div className="loading-spinner" />
+    </div>
+  );
+};
+
+export default Spinner;
