+++ code/src/index.css
@@ -1,5 +1,12 @@
+* {
+  margin: 0px;
+  padding: 0;
+  box-sizing: border-box;
+}
+
 body {
-  margin: 20px;
+  font-size: 12px;
+  font-weight: bold;
   font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
     "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
     sans-serif;
@@ -11,20 +18,86 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
     monospace;
 }
+
+.container {
+  margin: 0 auto;
+  margin-left: auto;
+  margin-right: auto;
+  width: 90%;
+}
+
+.heading-one {
+  text-align: center;
+  margin: 20px 0;
+  text-transform: capitalize;
+}
+/* Item component */
 .date {
   font-style: italic;
 }
+
+/* Form component */
+.form {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: 100%;
+}
+.new {
+  background-color: #f2f0f0;
+}
+
+.thought-container {
+  border: 1px solid #000;
+  padding: 15px;
+  margin: 0 auto;
+  margin-bottom: 30px;
+  box-shadow: 6px 6px;
+  width: 100%;
+  max-width: 540px;
+}
+.text-area {
+  display: block;
+  resize: none;
+  font-family: "JetBrains Mono", monospace;
+  /* font-family: var(--font); */
+  font-size: 12px;
+  font-weight: 700;
+  margin: 10px 0;
+  padding: 10px;
+  border-color: #e0dfdf;
+  width: 90%;
+}
+.thought-btn {
+  background-color: #ffadad;
+  border-color: transparent;
+  border-radius: 50px;
+  padding: 10px;
+  width: -webkit-fit-content;
+  width: -moz-fit-content;
+  width: fit-content;
+  font-family: "JetBrains Mono", monospace;
+  /* font-family: var(--font); */
+  font-size: 12px;
+  font-weight: 700;
+  color: #000;
+  transition: all 0.2s ease-in-out;
+}
+.thought-btn:hover {
+  transform: scale(1.1);
+}
+
+/* Loading component */
 .loading-overlay {
   position: absolute;
   top: 0;
-  bottom:0;
+  bottom: 0;
   left: 0;
   right: 0;
   background-color: #fff;
   display: flex;
   align-items: center;
   justify-content: center;
-
 }
 
 .loading-spinner {
